Core: Refactor savestates
