+++ src/core/serialize.c
@@ -5,7 +5,25 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 #include "serialize.h"
 
+#include "core/core.h"
+#include "core/cheats.h"
+#include "core/sync.h"
+#include "util/memory.h"
 #include "util/vfs.h"
+
+#ifdef USE_PNG
+#include "util/png-io.h"
+#include <png.h>
+#include <zlib.h>
+#endif
+
+mLOG_DEFINE_CATEGORY(SAVESTATE, "Savestate");
+
+struct mBundledState {
+	size_t stateSize;
+	void* state;
+	struct mStateExtdata* extdata;
+};
 
 struct mStateExtdataHeader {
 	uint32_t tag;
@@ -129,3 +147,296 @@ bool mStateExtdataDeserialize(struct mStateExtdata* extdata, struct VFile* vf) {
 	};
 	return true;
 }
+
+
+
+
+
+
+
+#ifdef USE_PNG
+static bool _savePNGState(struct mCore* core, struct VFile* vf, struct mStateExtdata* extdata) {
+	size_t stride;
+	color_t* pixels = 0;
+
+	core->getVideoBuffer(core, &pixels, &stride);
+	if (!pixels) {
+		return false;
+	}
+
+	size_t stateSize = core->stateSize(core);
+	void* state = anonymousMemoryMap(stateSize);
+	if (!state) {
+		return false;
+	}
+	core->saveState(core, state);
+
+	uLongf len = compressBound(stateSize);
+	void* buffer = malloc(len);
+	if (!buffer) {
+		mappedMemoryFree(state, stateSize);
+		return false;
+	}
+	compress(buffer, &len, (const Bytef*) state, stateSize);
+	mappedMemoryFree(state, stateSize);
+
+	unsigned width, height;
+	core->desiredVideoDimensions(core, &width, &height);
+	png_structp png = PNGWriteOpen(vf);
+	png_infop info = PNGWriteHeader(png, width, height);
+	if (!png || !info) {
+		PNGWriteClose(png, info);
+		free(buffer);
+		return false;
+	}
+	PNGWritePixels(png, width, height, stride, pixels);
+	PNGWriteCustomChunk(png, "gbAs", len, buffer);
+	if (extdata) {
+		uint32_t i;
+		for (i = 1; i < EXTDATA_MAX; ++i) {
+			if (!extdata->data[i].data) {
+				continue;
+			}
+			uLongf len = compressBound(extdata->data[i].size) + sizeof(uint32_t) * 2;
+			uint32_t* data = malloc(len);
+			if (!data) {
+				continue;
+			}
+			STORE_32LE(i, 0, data);
+			STORE_32LE(extdata->data[i].size, sizeof(uint32_t), data);
+			compress((Bytef*) (data + 2), &len, extdata->data[i].data, extdata->data[i].size);
+			PNGWriteCustomChunk(png, "gbAx", len + sizeof(uint32_t) * 2, data);
+			free(data);
+		}
+	}
+	PNGWriteClose(png, info);
+	free(buffer);
+	return true;
+}
+
+static int _loadPNGChunkHandler(png_structp png, png_unknown_chunkp chunk) {
+	struct mBundledState* bundle = png_get_user_chunk_ptr(png);
+	if (!bundle) {
+		return 0;
+	}
+	if (!strcmp((const char*) chunk->name, "gbAs")) {
+		void* state = bundle->state;
+		if (!state) {
+			return 0;
+		}
+		uLongf len = bundle->stateSize;
+		uncompress((Bytef*) state, &len, chunk->data, chunk->size);
+		return 1;
+	}
+	if (!strcmp((const char*) chunk->name, "gbAx")) {
+		struct mStateExtdata* extdata = bundle->extdata;
+		if (!extdata) {
+			return 0;
+		}
+		struct mStateExtdataItem item;
+		if (chunk->size < sizeof(uint32_t) * 2) {
+			return 0;
+		}
+		uint32_t tag;
+		LOAD_32LE(tag, 0, chunk->data);
+		LOAD_32LE(item.size, sizeof(uint32_t), chunk->data);
+		uLongf len = item.size;
+		if (item.size < 0 || tag == EXTDATA_NONE || tag >= EXTDATA_MAX) {
+			return 0;
+		}
+		item.data = malloc(item.size);
+		item.clean = free;
+		if (!item.data) {
+			return 0;
+		}
+		const uint8_t* data = chunk->data;
+		data += sizeof(uint32_t) * 2;
+		uncompress((Bytef*) item.data, &len, data, chunk->size);
+		item.size = len;
+		mStateExtdataPut(extdata, tag, &item);
+		return 1;
+	}
+	return 0;
+}
+
+static void* _loadPNGState(struct mCore* core, struct VFile* vf, struct mStateExtdata* extdata) {
+	png_structp png = PNGReadOpen(vf, PNG_HEADER_BYTES);
+	png_infop info = png_create_info_struct(png);
+	png_infop end = png_create_info_struct(png);
+	if (!png || !info || !end) {
+		PNGReadClose(png, info, end);
+		return false;
+	}
+	unsigned width, height;
+	core->desiredVideoDimensions(core, &width, &height);
+	uint32_t* pixels = malloc(width * height * 4);
+	if (!pixels) {
+		PNGReadClose(png, info, end);
+		return false;
+	}
+
+	size_t stateSize = core->stateSize(core);
+	void* state = anonymousMemoryMap(stateSize);
+	struct mBundledState bundle = {
+		.stateSize = stateSize,
+		.state = state,
+		.extdata = extdata
+	};
+
+	PNGInstallChunkHandler(png, &bundle, _loadPNGChunkHandler, "gbAs gbAx");
+	bool success = PNGReadHeader(png, info);
+	success = success && PNGReadPixels(png, info, pixels, width, height, width);
+	success = success && PNGReadFooter(png, end);
+	PNGReadClose(png, info, end);
+
+	if (success) {
+		struct mStateExtdataItem item = {
+			.size = width * height * 4,
+			.data = pixels,
+			.clean = free
+		};
+		mStateExtdataPut(extdata, EXTDATA_SCREENSHOT, &item);
+	} else {
+		free(pixels);
+		mappedMemoryFree(state, stateSize);
+		return 0;
+	}
+	return state;
+}
+#endif
+
+bool mCoreSaveStateNamed(struct mCore* core, struct VFile* vf, int flags) {
+	struct mStateExtdata extdata;
+	mStateExtdataInit(&extdata);
+	size_t stateSize = core->stateSize(core);
+	if (flags & SAVESTATE_SAVEDATA) {
+	/*	// TODO: A better way to do this would be nice
+		void* sram = malloc(SIZE_CART_FLASH1M);
+		struct VFile* svf = VFileFromMemory(sram, SIZE_CART_FLASH1M);
+		if (GBASavedataClone(&gba->memory.savedata, svf)) {
+			struct mStateExtdataItem item = {
+				.size = svf->seek(svf, 0, SEEK_CUR),
+				.data = sram,
+				.clean = free
+			};
+			mStateExtdataPut(&extdata, EXTDATA_SAVEDATA, &item);
+		} else {
+			free(sram);
+		}
+		svf->close(svf);*/
+	}
+	struct VFile* cheatVf = 0;
+	struct mCheatDevice* device;
+	if (flags & SAVESTATE_CHEATS && (device = core->cheatDevice(core))) {
+		cheatVf = VFileMemChunk(0, 0);
+		if (cheatVf) {
+			mCheatSaveFile(device, cheatVf);
+			struct mStateExtdataItem item = {
+				.size = cheatVf->size(cheatVf),
+				.data = cheatVf->map(cheatVf, cheatVf->size(cheatVf), MAP_READ),
+				.clean = 0
+			};
+			mStateExtdataPut(&extdata, EXTDATA_CHEATS, &item);
+		}
+	}
+#ifdef USE_PNG
+	if (!(flags & SAVESTATE_SCREENSHOT)) {
+#else
+	UNUSED(flags);
+#endif
+		vf->truncate(vf, stateSize);
+		struct GBASerializedState* state = vf->map(vf, stateSize, MAP_WRITE);
+		if (!state) {
+			mStateExtdataDeinit(&extdata);
+			if (cheatVf) {
+				cheatVf->close(cheatVf);
+			}
+			return false;
+		}
+		core->saveState(core, state);
+		vf->unmap(vf, state, stateSize);
+		vf->seek(vf, stateSize, SEEK_SET);
+		mStateExtdataSerialize(&extdata, vf);
+		mStateExtdataDeinit(&extdata);
+		if (cheatVf) {
+			cheatVf->close(cheatVf);
+		}
+		return true;
+#ifdef USE_PNG
+	}
+	else {
+		bool success = _savePNGState(core, vf, &extdata);
+		mStateExtdataDeinit(&extdata);
+		return success;
+	}
+#endif
+	mStateExtdataDeinit(&extdata);
+	return false;
+}
+
+void* mCoreExtractState(struct mCore* core, struct VFile* vf, struct mStateExtdata* extdata) {
+#ifdef USE_PNG
+	if (isPNG(vf)) {
+		return _loadPNGState(core, vf, extdata);
+	}
+#endif
+	ssize_t stateSize = core->stateSize(core);
+	vf->seek(vf, 0, SEEK_SET);
+	if (vf->size(vf) < stateSize) {
+		return false;
+	}
+	void* state = anonymousMemoryMap(stateSize);
+	if (vf->read(vf, state, stateSize) != stateSize) {
+		mappedMemoryFree(state, stateSize);
+		return 0;
+	}
+	if (extdata) {
+		mStateExtdataDeserialize(extdata, vf);
+	}
+	return state;
+}
+
+bool mCoreLoadStateNamed(struct mCore* core, struct VFile* vf, int flags) {
+	struct mStateExtdata extdata;
+	mStateExtdataInit(&extdata);
+	void* state = mCoreExtractState(core, vf, &extdata);
+	if (!state) {
+		return false;
+	}
+	bool success = core->loadState(core, state);
+	mappedMemoryFree(state, core->stateSize(core));
+
+	unsigned width, height;
+	core->desiredVideoDimensions(core, &width, &height);
+
+	struct mStateExtdataItem item;
+	if (flags & SAVESTATE_SCREENSHOT && mStateExtdataGet(&extdata, EXTDATA_SCREENSHOT, &item)) {
+		if (item.size >= (int) (width * height) * 4) {
+			/*gba->video.renderer->putPixels(gba->video.renderer, width, item.data);
+			mCoreSyncForceFrame(core->sync);*/
+		} else {
+			mLOG(SAVESTATE, WARN, "Savestate includes invalid screenshot");
+		}
+	}
+	if (flags & SAVESTATE_SAVEDATA && mStateExtdataGet(&extdata, EXTDATA_SAVEDATA, &item)) {
+		/*struct VFile* svf = VFileFromMemory(item.data, item.size);
+		GBASavedataLoad(&gba->memory.savedata, svf);
+		if (svf) {
+			svf->close(svf);
+		}*/
+	}
+	struct mCheatDevice* device;
+	if (flags & SAVESTATE_CHEATS && (device = core->cheatDevice(core)) && mStateExtdataGet(&extdata, EXTDATA_CHEATS, &item)) {
+		if (item.size) {
+			struct VFile* svf = VFileFromMemory(item.data, item.size);
+			if (svf) {
+				mCheatDeviceClear(device);
+				mCheatParseFile(device, svf);
+				svf->close(svf);
+			}
+		}
+	}
+	mStateExtdataDeinit(&extdata);
+	return success;
+}
+
