+++ src/gba/rr/rr.c
@@ -30,7 +30,7 @@ void GBARRInitRecord(struct GBA* gba) {
 
 	if (gba->rr->initFrom & INIT_FROM_SAVESTATE) {
 		struct VFile* vf = gba->rr->openSavestate(gba->rr, O_TRUNC | O_CREAT | O_RDWR);
-		GBASaveStateNamed(gba, vf, SAVESTATE_SAVEDATA);
+		//GBASaveStateNamed(gba, vf, SAVESTATE_SAVEDATA);
 		vf->close(vf);
 	} else {
 		ARMReset(gba->cpu);
@@ -54,7 +54,7 @@ void GBARRInitPlay(struct GBA* gba) {
 
 	if (gba->rr->initFrom & INIT_FROM_SAVESTATE) {
 		struct VFile* vf = gba->rr->openSavestate(gba->rr, O_RDONLY);
-		GBALoadStateNamed(gba, vf, SAVESTATE_SCREENSHOT | SAVESTATE_SAVEDATA);
+		//GBALoadStateNamed(gba, vf, SAVESTATE_SCREENSHOT | SAVESTATE_SAVEDATA);
 		vf->close(vf);
 	} else {
 		ARMReset(gba->cpu);
