+++ src/gba/serialize.c
@@ -6,7 +6,6 @@
 #include "serialize.h"
 
 #include "core/serialize.h"
-#include "core/sync.h"
 #include "gba/audio.h"
 #include "gba/cheats.h"
 #include "gba/io.h"
@@ -21,12 +20,6 @@
 #include <time.h>
 #else
 #include <sys/time.h>
-#endif
-
-#ifdef USE_PNG
-#include "util/png-io.h"
-#include <png.h>
-#include <zlib.h>
 #endif
 
 const uint32_t GBA_SAVESTATE_MAGIC = 0x01000000;
@@ -213,278 +206,6 @@ bool GBADeserialize(struct GBA* gba, const struct GBASerializedState* state) {
 		gba->rr->stateLoaded(gba->rr, state);
 	}
 	return true;
-}
-
-#ifdef USE_PNG
-static bool _savePNGState(struct GBA* gba, struct VFile* vf, struct mStateExtdata* extdata) {
-	unsigned stride;
-	const void* pixels = 0;
-	gba->video.renderer->getPixels(gba->video.renderer, &stride, &pixels);
-	if (!pixels) {
-		return false;
-	}
-
-	struct GBASerializedState* state = GBAAllocateState();
-	if (!state) {
-		return false;
-	}
-	GBASerialize(gba, state);
-	uLongf len = compressBound(sizeof(*state));
-	void* buffer = malloc(len);
-	if (!buffer) {
-		GBADeallocateState(state);
-		return false;
-	}
-	compress(buffer, &len, (const Bytef*) state, sizeof(*state));
-	GBADeallocateState(state);
-
-	png_structp png = PNGWriteOpen(vf);
-	png_infop info = PNGWriteHeader(png, VIDEO_HORIZONTAL_PIXELS, VIDEO_VERTICAL_PIXELS);
-	if (!png || !info) {
-		PNGWriteClose(png, info);
-		free(buffer);
-		return false;
-	}
-	PNGWritePixels(png, VIDEO_HORIZONTAL_PIXELS, VIDEO_VERTICAL_PIXELS, stride, pixels);
-	PNGWriteCustomChunk(png, "gbAs", len, buffer);
-	if (extdata) {
-		uint32_t i;
-		for (i = 1; i < EXTDATA_MAX; ++i) {
-			if (!extdata->data[i].data) {
-				continue;
-			}
-			uLongf len = compressBound(extdata->data[i].size) + sizeof(uint32_t) * 2;
-			uint32_t* data = malloc(len);
-			if (!data) {
-				continue;
-			}
-			STORE_32(i, 0, data);
-			STORE_32(extdata->data[i].size, sizeof(uint32_t), data);
-			compress((Bytef*) (data + 2), &len, extdata->data[i].data, extdata->data[i].size);
-			PNGWriteCustomChunk(png, "gbAx", len + sizeof(uint32_t) * 2, data);
-			free(data);
-		}
-	}
-	PNGWriteClose(png, info);
-	free(buffer);
-	return true;
-}
-
-static int _loadPNGChunkHandler(png_structp png, png_unknown_chunkp chunk) {
-	struct GBABundledState* bundle = png_get_user_chunk_ptr(png);
-	if (!bundle) {
-		return 0;
-	}
-	if (!strcmp((const char*) chunk->name, "gbAs")) {
-		struct GBASerializedState* state = bundle->state;
-		if (!state) {
-			return 0;
-		}
-		uLongf len = sizeof(*state);
-		uncompress((Bytef*) state, &len, chunk->data, chunk->size);
-		return 1;
-	}
-	if (!strcmp((const char*) chunk->name, "gbAx")) {
-		struct mStateExtdata* extdata = bundle->extdata;
-		if (!extdata) {
-			return 0;
-		}
-		struct mStateExtdataItem item;
-		if (chunk->size < sizeof(uint32_t) * 2) {
-			return 0;
-		}
-		uint32_t tag;
-		LOAD_32(tag, 0, chunk->data);
-		LOAD_32(item.size, sizeof(uint32_t), chunk->data);
-		uLongf len = item.size;
-		if (item.size < 0 || tag == EXTDATA_NONE || tag >= EXTDATA_MAX) {
-			return 0;
-		}
-		item.data = malloc(item.size);
-		item.clean = free;
-		if (!item.data) {
-			return 0;
-		}
-		const uint8_t* data = chunk->data;
-		data += sizeof(uint32_t) * 2;
-		uncompress((Bytef*) item.data, &len, data, chunk->size);
-		item.size = len;
-		mStateExtdataPut(extdata, tag, &item);
-		return 1;
-	}
-	return 0;
-}
-
-static struct GBASerializedState* _loadPNGState(struct VFile* vf, struct mStateExtdata* extdata) {
-	png_structp png = PNGReadOpen(vf, PNG_HEADER_BYTES);
-	png_infop info = png_create_info_struct(png);
-	png_infop end = png_create_info_struct(png);
-	if (!png || !info || !end) {
-		PNGReadClose(png, info, end);
-		return false;
-	}
-	uint32_t* pixels = malloc(VIDEO_HORIZONTAL_PIXELS * VIDEO_VERTICAL_PIXELS * 4);
-	if (!pixels) {
-		PNGReadClose(png, info, end);
-		return false;
-	}
-
-	struct GBASerializedState* state = GBAAllocateState();
-	struct GBABundledState bundle = {
-		.state = state,
-		.extdata = extdata
-	};
-
-	PNGInstallChunkHandler(png, &bundle, _loadPNGChunkHandler, "gbAs gbAx");
-	bool success = PNGReadHeader(png, info);
-	success = success && PNGReadPixels(png, info, pixels, VIDEO_HORIZONTAL_PIXELS, VIDEO_VERTICAL_PIXELS, VIDEO_HORIZONTAL_PIXELS);
-	success = success && PNGReadFooter(png, end);
-	PNGReadClose(png, info, end);
-
-	if (success) {
-		struct mStateExtdataItem item = {
-			.size = VIDEO_HORIZONTAL_PIXELS * VIDEO_VERTICAL_PIXELS * 4,
-			.data = pixels,
-			.clean = free
-		};
-		mStateExtdataPut(extdata, EXTDATA_SCREENSHOT, &item);
-	} else {
-		free(pixels);
-		GBADeallocateState(state);
-		return 0;
-	}
-	return state;
-}
-#endif
-
-bool GBASaveStateNamed(struct GBA* gba, struct VFile* vf, int flags) {
-	struct mStateExtdata extdata;
-	mStateExtdataInit(&extdata);
-	if (flags & SAVESTATE_SAVEDATA) {
-		// TODO: A better way to do this would be nice
-		void* sram = malloc(SIZE_CART_FLASH1M);
-		struct VFile* svf = VFileFromMemory(sram, SIZE_CART_FLASH1M);
-		if (GBASavedataClone(&gba->memory.savedata, svf)) {
-			struct mStateExtdataItem item = {
-				.size = svf->seek(svf, 0, SEEK_CUR),
-				.data = sram,
-				.clean = free
-			};
-			mStateExtdataPut(&extdata, EXTDATA_SAVEDATA, &item);
-		} else {
-			free(sram);
-		}
-		svf->close(svf);
-	}
-	struct VFile* cheatVf = 0;
-	if (flags & SAVESTATE_CHEATS && gba->cpu->components && gba->cpu->components[CPU_COMPONENT_CHEAT_DEVICE]) {
-		struct mCheatDevice* device = (struct mCheatDevice*) gba->cpu->components[CPU_COMPONENT_CHEAT_DEVICE];
-		cheatVf = VFileMemChunk(0, 0);
-		if (cheatVf) {
-			mCheatSaveFile(device, cheatVf);
-			struct mStateExtdataItem item = {
-				.size = cheatVf->size(cheatVf),
-				.data = cheatVf->map(cheatVf, cheatVf->size(cheatVf), MAP_READ),
-				.clean = 0
-			};
-			mStateExtdataPut(&extdata, EXTDATA_CHEATS, &item);
-		}
-	};
-#ifdef USE_PNG
-	if (!(flags & SAVESTATE_SCREENSHOT)) {
-#else
-	UNUSED(flags);
-#endif
-		vf->truncate(vf, sizeof(struct GBASerializedState));
-		struct GBASerializedState* state = vf->map(vf, sizeof(struct GBASerializedState), MAP_WRITE);
-		if (!state) {
-			mStateExtdataDeinit(&extdata);
-			if (cheatVf) {
-				cheatVf->close(cheatVf);
-			}
-			return false;
-		}
-		GBASerialize(gba, state);
-		vf->unmap(vf, state, sizeof(struct GBASerializedState));
-		vf->seek(vf, sizeof(struct GBASerializedState), SEEK_SET);
-		mStateExtdataSerialize(&extdata, vf);
-		mStateExtdataDeinit(&extdata);
-		if (cheatVf) {
-			cheatVf->close(cheatVf);
-		}
-		return true;
-#ifdef USE_PNG
-	}
-	else {
-		bool success = _savePNGState(gba, vf, &extdata);
-		mStateExtdataDeinit(&extdata);
-		return success;
-	}
-#endif
-	mStateExtdataDeinit(&extdata);
-	return false;
-}
-
-struct GBASerializedState* GBAExtractState(struct VFile* vf, struct mStateExtdata* extdata) {
-#ifdef USE_PNG
-	if (isPNG(vf)) {
-		return _loadPNGState(vf, extdata);
-	}
-#endif
-	vf->seek(vf, 0, SEEK_SET);
-	if (vf->size(vf) < (ssize_t) sizeof(struct GBASerializedState)) {
-		return false;
-	}
-	struct GBASerializedState* state = GBAAllocateState();
-	if (vf->read(vf, state, sizeof(*state)) != sizeof(*state)) {
-		GBADeallocateState(state);
-		return 0;
-	}
-	if (extdata) {
-		mStateExtdataDeserialize(extdata, vf);
-	}
-	return state;
-}
-
-bool GBALoadStateNamed(struct GBA* gba, struct VFile* vf, int flags) {
-	struct mStateExtdata extdata;
-	mStateExtdataInit(&extdata);
-	struct GBASerializedState* state = GBAExtractState(vf, &extdata);
-	if (!state) {
-		return false;
-	}
-	bool success = GBADeserialize(gba, state);
-	GBADeallocateState(state);
-
-	struct mStateExtdataItem item;
-	if (flags & SAVESTATE_SCREENSHOT && mStateExtdataGet(&extdata, EXTDATA_SCREENSHOT, &item)) {
-		if (item.size >= VIDEO_HORIZONTAL_PIXELS * VIDEO_VERTICAL_PIXELS * 4) {
-			gba->video.renderer->putPixels(gba->video.renderer, VIDEO_HORIZONTAL_PIXELS, item.data);
-			mCoreSyncForceFrame(gba->sync);
-		} else {
-			mLOG(GBA_STATE, WARN, "Savestate includes invalid screenshot");
-		}
-	}
-	if (flags & SAVESTATE_SAVEDATA && mStateExtdataGet(&extdata, EXTDATA_SAVEDATA, &item)) {
-		struct VFile* svf = VFileFromMemory(item.data, item.size);
-		GBASavedataLoad(&gba->memory.savedata, svf);
-		if (svf) {
-			svf->close(svf);
-		}
-	}
-	if (flags & SAVESTATE_CHEATS && gba->cpu->components && gba->cpu->components[CPU_COMPONENT_CHEAT_DEVICE] && mStateExtdataGet(&extdata, EXTDATA_CHEATS, &item)) {
-		if (item.size) {
-			struct mCheatDevice* device = (struct mCheatDevice*) gba->cpu->components[CPU_COMPONENT_CHEAT_DEVICE];
-			struct VFile* svf = VFileFromMemory(item.data, item.size);
-			if (svf) {
-				mCheatDeviceClear(device);
-				mCheatParseFile(device, svf);
-				svf->close(svf);
-			}
-		}
-	}
-	mStateExtdataDeinit(&extdata);
-	return success;
 }
 
 struct GBASerializedState* GBAAllocateState(void) {
