+++ src/platform/test/fuzz-main.c
@@ -102,7 +102,7 @@ int main(int argc, char** argv) {
 	}
 	if (savestate) {
 		if (!savestateOverlay) {
-			core->loadState(core, savestate, 0);
+			mCoreLoadStateNamed(core, savestate, 0);
 		} else {
 			struct GBASerializedState* state = GBAAllocateState();
 			savestate->read(savestate, state, sizeof(*state));
